requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
chamfer==2.0.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
coloredlogs==14.0
comm==0.2.0
configparser==5.2.0
contourpy==1.1.1
crcmod==1.7
cryptacular==1.5.5
cryptography==41.0.7
cycler==0.12.1
deb-pkg-tools==8.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
descartes==1.1.0
docker==4.3.1
docker-pycreds==0.4.0
easydict==1.7
emd-ext==0.0.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
executor==23.1
fasteners==0.15
fastjsonschema==2.19.0
filelock==3.13.1
fire==0.4.0
flatbuffers==2.0
fonttools==4.45.1
fqdn==1.5.1
fsspec==2023.10.0
ftfy==6.1.3
fvcore==0.1.5.post20221221
gast==0.4.0
gedit==0.0.2
gitdb==4.0.9
GitPython==3.1.26
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.10.0
huggingface-hub==0.19.4
humanfriendly==8.2
hupper==1.10.2
idna==3.6
imagecorruptions==1.1.2
imageio==2.9.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
install==1.3.5
iopath==0.1.10
ipykernel==6.27.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==0.16.0
json-tricks==3.15.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-packaging==0.11.1
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==12.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmcv==0.6.2
mmcv-full==1.4.0
model-index==0.1.11
monotonic==1.5
motmetrics==1.1.3
munch==2.5.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
ninja==1.10.0.post2
nose==1.3.7
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.22.1
oauthlib==3.1.1
objgraph==3.5.0
open3d==0.13.0
open3d-python==0.3.0.0
opencv-contrib-python==4.5.1.48
opencv-python==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.32
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
PasteDeploy==2.1.1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==4.0.0
pointnet2-ops @ file:///fastscratch/zhiminc/Bridge3D/Bridge3D/Pointnet2_PyTorch/pointnet2_ops_lib
pointops==0.0.0
portalocker==2.8.2
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.41
property-manager==3.0
protobuf==3.20.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycocotools==2.0.2
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.17.2
pyparsing==3.1.1
pyquaternion==0.9.9
pyramid==1.10.5
pyramid-mailer==0.15.1
pytest-runner==5.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@f74fc450e87ce63cea19249a54ef0bc117e1a39b
pytz==2023.3.post1
PyWavelets==1.1.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.31.1
regex==2023.10.3
repoze.sendmail==4.4.1
requests==2.28.2
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.13.2
safetensors==0.4.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==1.5.3
Shapely==1.7.1
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.3.22
stack-data==0.6.3
subprocess32==3.5.4
tabulate==0.9.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.3.2
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.18.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.9.22
timm==0.4.5
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.7.2
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tornado==6.4
tqdm==4.65.2
traitlets==5.14.0
transaction==3.0.1
transformers==4.35.2
transforms3d==0.3.1
translationstring==1.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
velruse==1.1.1
venusian==3.0.0
verboselogs==1.7
wandb==0.12.9
waymo-open-dataset-tf-2-3-0==1.2.0.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.6.4
widgetsnbextension==4.0.9
wrapt==1.16.0
WTForms==2.3.3
wtforms-recaptcha==0.3.2
y-py==0.6.2
yacs==0.1.8
yapf==0.40.2
yaspin==2.1.0
ypy-websocket==0.8.4
zipp==3.17.0
zope.deprecation==4.4.0
zope.interface==5.2.0
zope.sqlalchemy==1.3
